Fix low-risk audit issues: precision loss, safety guards, diagnostics
- #19: Remove Float→Double conversion on matchPercent in DeltaBuilder;
  switch availableCash/manualBudgetAmount from putFloat to putString
  for lossless persistence with backward-compatible getFloat fallback
- #20: Add NaN/Infinity guard in persistAvailableCash()
- #21: Split deserialization from unchecked cast in mergeRecordIntoList,
  add Log.w diagnostic on type mismatch
- #24: Replace coerceIn with coerceAtLeast(0) in amortization deductions,
  add clarifying comment for elapsed==totalPeriods boundary
- #22, #23, #25: Documented as accepted design choices in audit report
- Update audit report with resolution status for all 25 items

+++ app/src/main/java/com/syncbudget/app/data/sync/DeltaBuilder.kt
@@ -112,7 +112,7 @@ object DeltaBuilder {
         if (settings.resetHour_clock > lastPushedClock) fields["resetHour"] = FieldDelta(settings.resetHour, settings.resetHour_clock)
         if (settings.familyTimezone_clock > lastPushedClock) fields["familyTimezone"] = FieldDelta(settings.familyTimezone, settings.familyTimezone_clock)
         if (settings.matchDays_clock > lastPushedClock) fields["matchDays"] = FieldDelta(settings.matchDays, settings.matchDays_clock)
-        if (settings.matchPercent_clock > lastPushedClock) fields["matchPercent"] = FieldDelta(settings.matchPercent.toDouble(), settings.matchPercent_clock)
+        if (settings.matchPercent_clock > lastPushedClock) fields["matchPercent"] = FieldDelta(settings.matchPercent, settings.matchPercent_clock)
         if (settings.matchDollar_clock > lastPushedClock) fields["matchDollar"] = FieldDelta(settings.matchDollar, settings.matchDollar_clock)
         if (settings.matchChars_clock > lastPushedClock) fields["matchChars"] = FieldDelta(settings.matchChars, settings.matchChars_clock)
         if (settings.showAttribution_clock > lastPushedClock) fields["showAttribution"] = FieldDelta(settings.showAttribution, settings.showAttribution_clock)
